minionlib changes for compatibility between ffxiv and eso

+++ ml_task.lua
@@ -14,6 +14,7 @@ ml_task.overwatch_elements = {}
 ml_task.breakUpdate = false
 ml_task.delayTime = 0
 ml_task.delayTimer = 0
+ml_task.preserveSubtasks = false
 
 -- These functions are NOT overwritten in derived tasks
 
@@ -26,6 +27,11 @@ end
 
 function ml_task:isValid()
 	return self.valid
+end
+
+function ml_task:Invalidate()
+	ml_debug(self.name.."->Invalidate(Fail)()")
+	self.valid = false
 end
 
 function ml_task:hasCompleted()
@@ -86,13 +92,19 @@ function ml_task:Update()
 		end
 		
 		if(self:ProcessOverWatch()) then
-			ml_debug(self.name.."->ProcessOverWatch executed an effect, breaking loop")
-			--process overwatch element requested to break update loop
-			--only delete subtask if we didn't just add it via our overwatch cne
-			if (self.subtask ~= nil and (currentSubtaskName ~= nil or self.subtask.name == currentSubtaskName)) then
-				self:DeleteSubTasks()
+			if (not self.preserveSubtasks) then
+				ml_debug(self.name.."->ProcessOverWatch executed an effect, breaking loop")
+				--process overwatch element requested to break update loop
+				--only delete subtask if we didn't just add it via our overwatch cne
+
+				if (self.subtask ~= nil and (currentSubtaskName ~= nil or self.subtask.name == currentSubtaskName)) then
+					self:DeleteSubTasks()
+				end
+				
+				break
+			else
+				self.preserveSubtasks = false
 			end
-			break
 		end
 		
         if ( self.subtask ~= nil ) then
@@ -151,14 +163,25 @@ function ml_task:ProcessOverWatch()
 end
 
 function ml_task:SetDelay(delayTimer)
-	if(	type(delayTimer) == "number" and
-		delayTimer > 0)
-	then
+	--ignore 0 case
+	if(	type(delayTimer) == "number" and delayTimer > 0) then
 		self.delayTime = ml_global_information.Now
 		self.delayTimer = delayTimer
-	else
-		ml_error("Invalid delaytimer input")
+	elseif(	type(delayTimer) == "number" and delayTimer < 0) then
+		ml_error("Invalid delaytimer input - negative number")
 	end
+end
+
+function ml_task:IsDelayed()
+	return TimeSince(self.delayTime) < self.delayTimer
+end
+
+function ml_task:GetDelay()
+	if(self:IsDelayed()) then
+		return self.delayTimer - TimeSince(self.delayTime)
+	end
+	
+	return nil
 end
 
 --These functions ARE overwritten in derived tasks
@@ -241,6 +264,28 @@ function ml_task:AddTaskCheckCEs()
 
     local ke_fail = ml_element:create( "TaskFail", c_fail, e_fail, ml_effect.priorities.interrupt )
 	self:add( ke_fail, self.process_elements)
+end
+
+function ml_task:AddOverwatchTaskCheckCEs()
+    --Add complete check
+    local c_complete = inheritsFrom(ml_cause)
+    function c_complete:evaluate() return ml_task_hub:CurrentTask():task_complete_eval() end
+    
+    local e_complete = inheritsFrom(ml_effect)
+    function e_complete:execute() ml_task_hub:CurrentTask():task_complete_execute() end
+
+    local ke_complete = ml_element:create( "TaskComplete", c_complete, e_complete, ml_effect.priorities.interrupt )
+	self:add( ke_complete, self.overwatch_elements)
+    
+    --Add fail check
+    local c_fail = inheritsFrom(ml_cause)
+    function c_fail:evaluate() return ml_task_hub:CurrentTask():task_fail_eval() end
+    
+    local e_fail = inheritsFrom(ml_effect)
+    function e_fail:execute() ml_task_hub:CurrentTask():task_fail_execute() end
+
+    local ke_fail = ml_element:create( "TaskFail", c_fail, e_fail, ml_effect.priorities.interrupt )
+	self:add( ke_fail, self.overwatch_elements)
 end
 
 function ml_task.Create()
